client re-write
about to add lots more complexity, so breaking into partial classes
